separated reducers to make it clearer to read
