+++ client/store/index.js
@@ -3,9 +3,10 @@ import {createLogger} from 'redux-logger'
 import thunkMiddleware from 'redux-thunk'
 import {composeWithDevTools} from 'redux-devtools-extension'
 import user from './user'
-import logs from './logs'
+import logs from './allLogsReducer'
+import singleLog from './singleLogReducer'
 
-const reducer = combineReducers({user, logs})
+const reducer = combineReducers({user, logs, singleLog})
 const middleware = composeWithDevTools(
   applyMiddleware(thunkMiddleware, createLogger({collapsed: true}))
 )
